Modified parameters for Invoke-RpCommandObject & Added new calls for configuration commands in Set-RpRunspaceEvents
**Note on Calling Get-VmsCameraReport**

When using the default configuration commands, there was some resistance in returning results from the command object. This issue may stem from the underlying functionality of Get-VmsCameraReport when working with a dedicated, reusable, isolated runspace.

This problem was present in a previous version of MIPSDK, but it may have been resolved by now. A similar issue also occurred with Show-RpCamera, which was resolved after addressing a memory leak. If this situation persists, the command object may need to be modified to return results using a standard runspace. This change would allow configuration commands to be executed for this report.